requirements.txt
langchain
langgraph
langsmith
langchain-community
langchain-google-genai
google-generativeai
langchain-core
pypdf
python-dotenv
openai
pandas
numpy
gunicorn
uvicorn
fastapi
python-multipart
pydantic
scikit-learn
scipy
spacy
streamlit
transformers
sentence-transformers
diffusers
torch
accelerate
BeautifulSoup4
PyPDF2
requests
langchain-google-genai
chromadb
langchain_chroma
pymupdf
langchain-huggingface
crawl4ai
python-jobspy